# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ovumcy API
  version: 1.0.0
  summary: HTTP API surface of the ovumcy-web self-hosted deployment.
  description: |
    This document specifies the **current** state of the JSON-emitting HTTP
    surface of `ovumcy-web`. It is the authoritative description of what the
    server actually serves today.

    Stability — per [CONTRIBUTING.md](../CONTRIBUTING.md#api-stability-contract):

    - `/api/v1/*` is the canonical, stable HTTP surface. Wrappers and external
      integrations should target this prefix exclusively.
    - Field additions are non-breaking and may ship in any minor release.
    - Field renames, removals, status code changes, route moves, and error key
      changes are breaking and require a new major version (`/api/v2/*`).
    - The export payload (`/api/v1/exports/*`) follows the separate stability
      contract documented in [docs/export.md](export.md).

    Content negotiation — every endpoint serves both JSON and HTML/HTMX from the
    same handler. This document covers **only** the JSON contract. Clients must
    send `Accept: application/json` to opt into the JSON shape. Without that
    header the server returns HTML or `303` redirects suitable for browser
    navigation, and the corresponding response is not part of this contract.

    Methods — every operation served on `GET` is also served on `HEAD` by the
    same handler chain: same status, same headers, no body. Those mirrors are
...
    documented per path below.

...

...

...

...
servers:
...
security:
  - sessionCookie: []
    csrfToken: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ovumcy_auth
      description: |
        Sealed session cookie issued by `POST /api/v1/sessions` (or by completing
        the OIDC sign-in flow). Verified against `users.auth_session_version` on
        every request. See [SECURITY.md](../SECURITY.md#cookies).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token validated against the `ovumcy_csrf` cookie.
        Browser clients usually submit it via the `csrf_token` form field
        instead of the header. Only pure-GET requests and the
        `/auth/oidc/callback` provider form-post path are exempt; every other
        state-changing request — including register, login, and password-reset
        — requires the token. Programmatic clients read it from the
        `ovumcy_csrf` cookie set on any GET (for example `GET /login`) and echo
        it in the `X-CSRF-Token` header or `csrf_token` form field. See
        [CSRF middleware configuration](../cmd/ovumcy/main.go).
...
paths:
...
  /lang:
    post:
      tags: [meta]
      summary: Set UI language preference.
      description: |
        Writes the `ovumcy_lang` cookie for the chosen language. Public, no session required.
        When the request does carry a valid session, the language is also stored on the account
        (`users.interface_language`, the same column `PATCH /api/v1/users/current/interface`
        writes), so a later session issue re-serves the switched language instead of reverting
        the cookie to the stored one. The response is the same either way, unless that account
        write fails (`500`).

        This path is outside `/api`, so the `/api` rate limit does not reach it; it carries a
        per-IP limit of its own, sized like the `/api` budget, and answers `429` once it is spent.
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [lang, csrf_token]
              properties:
                lang:
                  type: string
                  description: |
                    Form field name is `lang` (the handler reads `c.FormValue("lang")`) — distinct
                    from the `language` JSON field on `PATCH /api/v1/users/current/interface`. The
                    shipped codes are `en`, `ru`, `es`, `fr`, `de` and `it`. The value is not
                    validated against that list: a blank value is refused with `400`, and any
                    other code the server does not ship falls back to the instance's default
                    language.
                csrf_token: { type: string }
                next:
                  type: string
                  description: Optional same-origin path to return to; anything else falls back to `/`.
      responses:
        '200':
          description: HTMX-style success.
        '303':
          description: |
            Redirect to the form's `next` path, or to `/` when `next` is absent or is not a
            same-origin path.
        '400':
          description: |
            `lang` is missing or blank. The handler raises the transport-level `400`, so the
            refusal carries the shared `bad_request` key rather than a key of this route's own.
            A non-blank code the server does not ship is not refused: it falls back to the
            instance's default language and answers like a success.

            A caller that asks for JSON gets the JSON envelope below: one whose `Accept`
            header names `application/json`, or one that sends its body as
            `Content-Type: application/json`, either one alone being enough. An HTMX request
            (`HX-Request: true`) outranks both and gets the localized status fragment as
            `text/html` instead, and so does a plain form submission that asks for neither —
            one sending `Accept: */*` or no `Accept` header at all included. That fragment is
            the whole page the browser shows, so it ends with a link back to the form's `next`
            path, or to `/` when `next` is absent or is not a same-origin path.
          content:
            text/html:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error: "bad_request"
                error_detail: { key: "bad_request", category: "validation", target: "global" }
        '403':
          description: |
            The CSRF check refused the request: no token; a token that does not match the
            `ovumcy_csrf` cookie, or one the server no longer holds (idle for over an hour, or
            issued before the server restarted — that answer also clears the cookie, so fetch a
            page again); an `Origin` whose scheme or host differs from the server's, a sibling
            subdomain included, or over HTTPS with no `Origin`, a missing or foreign `Referer`;
            or a malformed `Sec-Fetch-Site`. The refusal comes from the CSRF middleware before the
            handler runs, so no `ovumcy_lang` cookie is set and nothing is stored on the
            account. It carries the shared `forbidden` key rather than a key of this route's
            own.

            A caller that asks for JSON gets the JSON envelope below: one whose `Accept`
            header names `application/json`, or one that sends its body as
            `Content-Type: application/json`, either one alone being enough. An HTMX request
            (`HX-Request: true`) outranks both and gets the localized status fragment as
            `text/html` instead, and so does a plain form submission that asks for neither —
            one sending `Accept: */*` or no `Accept` header at all included. That fragment is
            the whole page the browser shows, so it ends with a link back to the form's `next`
            path, or to `/` when `next` is absent or is not a same-origin path.
          content:
            text/html:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error: "forbidden"
                error_detail: { key: "forbidden", category: "forbidden", target: "global" }
        '500':
          description: |
            The request carries a live owner session and a shipped language code, and storing
            that language on the account failed. The switch is reported as failed rather than
            kept as a cookie-only change: no `ovumcy_lang` cookie is set. A caller with no
            session, or one that sends a code the server does not ship, never reaches the
            account write and never gets this answer.

            Any other fault while answering the request, such as a recovered panic, is a `500`
            too, carrying the shared `internal_error` key.

            A caller that asks for JSON gets the JSON envelope below: one whose `Accept`
            header names `application/json`, or one that sends its body as
            `Content-Type: application/json`, either one alone being enough. An HTMX request
            (`HX-Request: true`) outranks both and gets the localized status fragment as
            `text/html` instead, and so does a plain form submission that asks for neither —
            one sending `Accept: */*` or no `Accept` header at all included. That fragment is
            the whole page the browser shows, so it ends with a link back to the form's `next`
            path, or to `/` when `next` is absent or is not a same-origin path.
          content:
            text/html:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error: "failed to update interface settings"
                error_detail: { key: "failed to update interface settings", category: "internal", target: "global" }
        '429':
          $ref: '#/components/responses/RateLimited'
          description: |
            Rate limit exceeded. Every answer carries the `Retry-After` header. A caller that
            asks for JSON gets the JSON envelope, with `retry_after_seconds`: one whose `Accept`
            header names `application/json`, or one that sends its body as
            `Content-Type: application/json`, either one alone being enough. An HTMX request
            (`HX-Request: true`) outranks both and gets the localized status fragment as
            `text/html` instead, with no `retry_after_seconds`, and so does the plain form
            submission (this route's primary client) that asks for neither.

  # =========================================================================
  # auth
  # =========================================================================
  /api/v1/users:
    post:
      tags: [auth]
      summary: Create a new owner account.
      description: |
        Returns the same response shape (a single sealed `ovumcy_register_pickup`
        cookie of fixed length, redirect to `/register/welcome`) for new and
        duplicate emails. See [SECURITY.md "Register enumeration"](../SECURITY.md#register-enumeration-pickup-cookie-follow-up-oracle-residual).
        Per-IP rate limit: default 8 / 15 min.
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: |
            Pickup cookie issued. Sent when `Accept: application/json` is set; the
            browser surface gets the `303` below instead. Identical for new and
            duplicate emails.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NextStepResponse' }
              example:
                ok: true
                next_step: "register_welcome"
                next_path: "/register/welcome"
        '303':
          description: Redirect to `/register/welcome` with the sealed pickup cookie. Shape is identical for new and duplicate emails.
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: |
            Registration is closed on this deployment (SSO-only) — carries this route's own key,
            `local sign-in unavailable`, target `auth_form`; a plain browser Accept flash-redirects
            to `/register` (see the `303` above) rather than answering here directly.

            A CSRF refusal (no token; a token that does not match the `ovumcy_csrf` cookie, or one
            the server no longer holds; a foreign `Origin`/`Referer`) is a separate cause reached
            before the handler runs, so it carries the shared `forbidden` key instead. A caller
            that asks for JSON gets the JSON envelope; a plain form submission that asks for
            neither gets the shared localized status fragment as `text/html` (WEB-84) — the same
            shape `POST /lang` documents above — instead of the raw envelope.
          content:
            text/html:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              examples:
                registrationClosed:
                  summary: Local sign-in disabled
                  value:
                    error: "local sign-in unavailable"
                    error_detail: { key: "local sign-in unavailable", category: "forbidden", target: "auth_form" }
                csrfDenied:
                  summary: CSRF refusal
                  value:
                    error: "forbidden"
                    error_detail: { key: "forbidden", category: "forbidden", target: "global" }
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/sessions/2fa-challenge:
    post:
      tags: [auth]
      summary: Submit the TOTP code in the 2FA challenge.
      description: |
        Requires the `ovumcy_totp_pending` cookie issued by a successful
        password login. Per-account TOTP rate limit: 5 attempts / 15 min.
        Replayed codes within the 30-second step are rejected with the same
        response shape as an invalid code.
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TOTPLoginRequest' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/TOTPLoginRequest' }
      responses:
        '200':
          description: 2FA verified; `ovumcy_auth` issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '303':
          description: Browser redirect to `/`.
        '401':
          description: Invalid code or pending session expired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '403':
          description: |
            A CSRF refusal (no token; a token that does not match the `ovumcy_csrf` cookie, or one
            the server no longer holds; a foreign `Origin`/`Referer`) reached before the handler
            runs, carrying the shared `forbidden` key. A caller that asks for JSON gets the JSON
            envelope below; a plain form submission that asks for neither gets the shared
            localized status fragment as `text/html` (WEB-84) — the same shape `POST /lang`
            documents above — instead of the raw envelope.
          content:
            text/html:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error: "forbidden"
                error_detail: { key: "forbidden", category: "forbidden", target: "global" }
        '429':
          $ref: '#/components/responses/RateLimited'
...